+++ src/Atoms/index.js
@@ -1,4 +1,4 @@
-export { default as BackToTop } from "./Buttons/BackToTop";
+// export { default as BackToTop } from "./Buttons/BackToTop";
 export { default as PageTitle } from "./PageTitle";
 export { default as Spinner } from "./LoadingSpinner";
 export { default as AlertMessage } from "./Messages/AlertMessage";
